REFACTOR moved notify onto its own file
